+++ packages/stream_chat/CHANGELOG.md
@@ -10,6 +10,7 @@
 - [[#890]](https://github.com/GetStream/stream-chat-flutter/pull/890) Fixed Reactions not updating on thread messages.
   Thanks [bstolinski](https://github.com/bstolinski).
 - [[#897]](https://github.com/GetStream/stream-chat-flutter/issues/897) Fixed error type mis-match in `AuthInterceptor`.
+- Fix `channelState.copyWith` with respect to pinnedMessages.
 
 ## 3.4.0