+++ packages/stream_chat/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Upcoming
+## 3.5.0
 
 ✅ Added